requirements.txt
paho-mqtt==1.2
configparser==3.5.0
pytz==2016.4